Run test suite in expected way
Turns out that the standard library test suite for the socket modules is run in a different fashion than normal unittest-based test suites. Particularly, they rely on execution of their `test_main` method, which only runs a select subset of tests. Using the default unittest test loader instead causes tests to run that potentially fail, as it happened here.

NB: locally the ssl tests still fail more, but I'm not 100% certain as to why. So I'm submitting this to rule out that it's an environmental issue with my machine.