Add missing files for `test_ssl.py`
This is a bit of an odd CL. Basically, the ssl module tests rely on a few things from the standard library, for example the `make_https_server` method. The trouble is now that `make_https_server` uses a request handler which resolves files relative to the location of `ssl_servers.py`. This has a few consequences, now that the ssl tests are run as part of this module, as opposed to the standard library test suite. Most notably, it requires changing the way that the helper method is imported, otherwise most of the standard library needs to be moved into this test suite. Additionally, because there is still a `test` subfolder expected to serve `keycert3.pem` from then the file is duplicated into that location.