Extract code relevant to CarbonIO/StacklessIO from blue
This is to be aware of expected APIs, and functionality that needs to be retained to some degree. The exact degree in which it needs to stay around is currently unknown. There is a chance that most of this code will be replaced by libuv's constructs. But until then, this is kept as reference material.