Remove separate thread for ticking libuv
Running a separate thread for ticking libuv causes interesting race conditions, for example, double callback invokations, which in case of the close callback cause a double delete.
Therefore, uv is now ticked in the same thread as the Python code. This mechanism was also in place for StacklessIO previously. However, the code is getting quite unreadable now, and needs some refactoring.