Refactor on_accept to be more in line with libuv documentation
Some tests fail due to ENOTCONN errors. An initial explanation for this was the implementation of the uv_accept call, which was different from the examples in the documentation. While this did not resolve the errors, the new implementation is more error proof and readable (Cleaning up resources using scope guards).