Stop libuv from reading data once the request amount was received
Initially, a read request was able to outlive the call to `recv()` because libuv was never told to stop reading data from the network. This worked fine to satisfy all the tests that were enabled until then, but now it is posing a headache with correctly tearing down the outstanding request object at the right moment in time. Simply deleting the request will cause us a use-after-free bug, because the delete closes the stackless channel. This in turn sends a `None` value and therefore wakes up the request that's currently waiting for data. However, this request is the one that was just deleted and had its destructor run. Therefore, all operations in the resumed function are undefined in their behaviour.