Rework receive buffer handling
The buffer now lives on the handle as opposed to on the request object. Keeping the request around after executing it was causing trouble, and we only really need to hold on to the unread data.
Tagged pointer pattern was getting untenable because we need to store the buffer and related book-keeping on the handle, so we swapped it out for using a struct.