Fix race condition between timeout and cleanup_uv_handle
It is possible that the uv_close_callback and the uv_timeout_callback for a request are invoked in the same uv tick. This would result in some interesting behaviors: When sending timeout errors over the channel, we would wake up the request which would keep processing and then clean up the channel. This would result in PyChannel_SendThrow returning an error since the channel got cleaned up from under its feet. This would result in our sendError function falsely believing that it had failed to send the error over the channel, and writing an unraisable error.