Improve translation of libuv errors to Python exceptions
Multiple instances happened where tests failed because the conversion of libuv error did not translate into an expected Python exception object. Examples include not returning the expected exception type, as well as not having the `errno` attribute on an exception. Since libuv error codes are the negated OS errno, they can be translated into the correct Python errno constant, which in turn can be used to create a Python exception via `PyErr_SetFromErrno`. This resolves all issues encountered with exception handling in the tests, and thus appears to be more in line with the way Python expects these exceptions to be.