Add timeout handling for send operations
Initially timeout handling for StreamSendRequest objects was skipped because libuv doesn't have a concept of timing out on send. However, Python expects this functionality to exist, as indicated by `SendfileUsingSendTest.testWithTimeoutTriggeredSend`. Another discovery was that libuv is so efficient at sending data, that the default timeout in the Python test suite was too high and would never be hit. Therefore, it got reduced by 2 orders of magnitude.