Limit `IPV6_ONLY` checks to `AF_INET6` sockets
This limitation sort of existed already in the Windows world, where Windows is fine with checking the `IPPROTO_IPV6` socket options on sockets with any kind of address family. macOS, on the other hand, is a lot more strict and fails with an `EINVAL` error unless the socket is `AF_INET6`. Therefore, the guard needs to be moved around the `getsockopt` call, as opposed to the `bind_flags` assignments.