Fix request lifetime issues
Requests were being cleaned up after the execute function completed, which caused issues if those requests had outstanding libuv callbacks.

Change interface to requests to be more uniform across the board with an 'execute()' method to perform the request action, and an 'onCallback()' method to handle the appropriate libuv callback.