Fix hanging sends when timeout occurs
Changing the lifetime of the requests to last beyond the scope of their execution revealed an issue with timeout handling:
When a timeout occurs it sends an exception over the channel. This unblocks any waiting party. However, any UV callback relating to the timed out request may not have completed yet. Now, if such a callback fires after a timeout and tries to send over the channel it will block indefinitely.
Therefore, add a flag to track if a timeout has occurred. If the timeout flag is set, the UV callback does not attempt to send over the channel.