Implement `socket.sendpacket()`
This follows the "headerless" CarbonIO packet format because that is also what the old StacklessIO/CarbonIO python exposed functions did. In StacklessIO these exposed function could, in theory, receive an additional header, but that was never used in the calling Python code. Primarily because the CarbonIO Python exposure did not support header data.