Fix issues when receiving packets on multiple tasklets
Calling recvpacketoob could result in a BlockingIOError when called from a tasklet while another tasklet is still waiting for a packet from that socket. Therefore, use a stackless channel to make sure only one tasklet can receive a packet at a time.

Once the channel had been introduced and the BlockingIOError fixed, it became clear that it is not safe to assume that the request stored on the socket is the one that the uv callback wants, as crashes started occurring where the callback for the read request started calling the onCallback method for the send request. To fix this, the receiveRequest member has been added. It should always point to the last receive request, which should be sufficient as only one read should be active at a time (or else we would get a BlockingIOError). Theoretically, this crash could happen for the sends as well, so we should look into that as a follow-up.

PLAT-5237 (Fix WinSock 10037 error in new socket module when used in client/proxy/server)