Use UV sockets with SSL
When wrapping a Python socket, the SSL module creates read/write BIOs around the file descriptor. This means, that it is not using the Python sockets as the underlying transport mechanism. This is unfortunate because the Python sockets with our customizations provide asynchronous IO. Therefore, take advantage of the OpenSSL BIO abstraction to provide BIOs that read/write through our python-wrapped UV sockets.