Introduce channel-per-request-object paradigm
Sharing a single channel for all operations on the socket to synchronize their operations works great when you are working with a single tasklet. However, when many tasklets use the socket at the same time, they can get confused if the operations do not complete in the exact same order that they were started. Therefore, each request maintains its own channel (except for the accept request, which uses a channel that lives on the socket and should be used only for the purpose of accepting connections).