Link against Blue instead of CcpCore
In order to use the carbon-core's logging functionality then it is unfortunately necessary to link against carbon-blue. This is due to some technical debt where carbon-core's logging functionality uses static variables to keep state around, which is conflicting with the static linking that happens as every library that links carbon-core gets its own version of the logging state.