Remove AF_UNIX constant
The standard library uses the presence of the AF_UNIX constant on the C++ socket module in order to decide which socket type to use when creating a socket pair. Since we don't currently have support for AF_UNIX sockets, I'm removing that constant to signal that this address family is not supported.