Fix skipping of tests only known to work on linux
The test was written to work on linux, but expects an SSLError to get raised on Windows, resulting in skipping the tests.

The comments say that "Typically WRONG_VERSION_NUMBER is the error", but with our modifications we get SSL_ERROR_EOF. Therefore, instead of the string regex, we skip the tests if the exception is an SSLError.