Make non-blocking send the default behavior for sockets
Blocking send was the default, because that was the default behavior in StacklessIO. However, non-blocking behavior was the default behavior in (old) CarbonIO and this is the behavior our server code expects. Therefore, update the default behavior to not blocking when sending data over a socket.

Update a couple of tests that expect blocking sends to time out or keep the sender alive until the receiver has received all the relevant data.