Enable blocking send for SocketConnectedTest client
This socket is often used to send data, and not setting it to blocking mode causes some test failures on my local machine. Instead of trying to pick and choose which specific tests to enable blocking sends for, it's simpler to set this on the test class level.

This also touches NetworkConnectionTest because it overrides the method used to set up the client socket in SocketConnectedTest