Fix hanging socket tests on Mac
A test would hang indefinitely because a thread would stop running, and the ScheduleManager for that thread would not get cleaned up. Then another thread would get created with the same thread ID, and chaos ensued. This turned out to be because the StreamSendRequest was not cleaning up a reference to the current tasklet.