Refactor large parts of request handling
Unfortunately a huge commit that's difficult to pick apart at this stage. It contains the following:

1. Correct use of `std::shared_ptr`

Previous usage of simply assigning `this` isn't a safe use of shared pointers. One has to inherit from `std::enable_shared_from_this` and then use `shared_from_this()`. This in turn only works when a shared pointer already exists, so all the requests are now shared pointer instances.

2. Tons of logging

Yes this is slow. But it's necessary for the time being in order to figure out where things go wrong. The general structure of the logging is such that we always log out the relevant uv_handle address, which makes searching and grouping in LogLite a lot simpler.

3. Enforce sequenced sends and receives

Otherwise, in a live environment we have seen tasklets stepping on each others toes when it comes to the payloads, leading to some odd corruption. Whether this is actually required for the send functionality is questionable, but I kept those changes in as there were some odd looking things in the logs.

4. Fix a potential use-after-free situation in `StreamSendRequest`

The underlying buffer for non-blocking sends may have been unused at the given point in time, so we're now memcpy'ing into a dedicated struct. Also, different callbacks are provided based on whether a send request is "blocking" or not.

5. Fix a potential race conditions around `HandleData::request` when `connect` and `send` interleave

This had potential to cast the wrong thing in the callback.

6. Miscellaneous

A few coding style fixes and missing default initializations which shouldn't have mattered, but make things a bit easier to read.