Fix copy from backing buffer into packet buffer
Three issues in this single line: First of all, it was always writing data to the start of the packet buffer. Of course that will break for packets larger than the default buffer size. Secondly, the size of the destination buffer was then incorrectly passed on; it needs to be reduced by the amount of bytes already written into the packet buffer. Last, but not least, the start offset into the socket buffer was wrongly calculated; `bufReadPos` only needs to be considered if the socket buffer isn't supposed to be copied in its entirety.