Call `uv_accept` from `StreamAcceptRequest`
Calling `uv_accept` from the `uv_listen` callback is too early; it will signal connecting clients that they are connected before the server-side has called `connect`. Therefore, the callback to `uv_listen` simply needs to signal on the incoming connection channel if someone is waiting, or alternately keep track of an incoming connection that can be accepted.