Fix handling of packets with OOB data
The initial Time Dilation packet caused the proxy to terminate the connection between SOL node and proxy, and subsequently proxy and client. This happened because "BlueNet ate the packet", e.g. the callback encountered a packet with OOB data but nothing else. However, carbon-io did not behave correctly in that scenario and returned `None` back up to Python, which is interpreted by machoNet as "something went horribly wrong, let's close the connection". Thus, this is now correctly handling packets that do not contain any real data by resetting the internal state of the request to be "we received nothing yet".

https://ccpgames.atlassian.net/browse/PLAT-6039