When receiving packet, eat malformed data instead of raising error
Old carbonIO would react to malformed packets by not attempting to parse any more packets from the chunk of received data. This means that any packet with a size field larger than the max packet size would disappear, and so would any potentially legit packets contained in the rest of the data. This change replicates that behavior.