Stop hanging reads when last data to receive already in buffer
When reading the last of a batch of data, requests would hang. This is because a read over the socket would be scheduled and blocked on, but there was no more data being sent over so that read would not complete. Therefore, if we have buffer in the data, just return that. The requested length is intended to be an upper bound for the amount of data we are prepared to receive, it should not be used to assume the amount of data that would be received.