Make _socket module link libuv as a shared library
This should have happened as part of the change exposing the uv loop. Sharing a UV loop when the library is linked statically makes absolutely no sense, since the consumer would have the library linked differently, resulting in different internal state for libuv. This resulted in strange behavior when attempting to share the uv_loop with carbon-db.