Integrate documentation generation using Doxygen and Sphinx
The generated documentation explains the purpose of carbon-io as well as how to use it. Additions made to the _socket and _ssl modules are documented, as are the additions made to the C API capsule. There is also a small section that explains how this interacts with other Carbon components.

Comments and docstrings have been updated in places in an attempt to make them more informative and readable.

Comment style has been updated to make comments show up in the generated documentation.

Docstrings have been formatted in a way that makes for prettier documentation.