Early out when failing to acquire send/receive
When closing a uv_handle, any outstanding read requests were getting their tasklets run before cleaning up the handle and continuing the dispatch loop. Blocking in the dispatch loop is a bad idea, since libuv expects that uv_run will not be called again while it is already running.

Therefore, sending Py_True to represent that permission to receive has been granted, and Py_False to mean that permission has not been granted and the handle has been cleaned up.

Note that even in the case where Py_True has been sent over the channel, this does not mean that uv_close has not been called between the sending tasklet placing the value on the channel and the receiving tasklet waking up, so just to be sure, we check the state of the uv_handle after waking up, even if it looks like everything is in order.