Prevent use-after-free when read requests get cancelled
When a read requests gets cancelled the cancel method would send the sentinel value over the channel to unblock the request tasklet, but that message did not contain any information about whether the request should expect the handle to be valid or not.

In the case where the request was cancelled and the handle was about to get cleaned up, the request would continue as if the handle was valid, but by the time its tasklet got around to running, the handle and handle data had become invalid memory.

Therefore, sending Py_True over the channel when data gets received and Py_False over the channel when the request is cancelled and the handle/handleData should not be accessed.