Fix test failure when receiving packets on a closed socket
The tests were failing because sending a packet would raise an OSError with errno ENOTCONN as a result of trying to read on a closed handle. Therefore, add a pointer to the handle to the RequestData object and invalidating it when we close the handle. This should prevent the use of invalid memory from requests accessing the handle or handleData.