Remove unnecessary booleans from RequestData
The reading and sending boolean flags are not necessary. Removing them and instead creating local shared pointers to keep the requests alive for the execute methods of the connect and accept request. That means we can use the pointers in RequestData to keep track of the currently sending and currently receiving requests exclusively, resulting in easier to understand code.