Prevent calling uv_read_stop with nullptr handle
When a handle gets closed, the receive requests cancel method gets called. The cancel methods attempt to call uv_read_stop with their uv handle,

Since they call handle() to get it, the handle is closing, they end up with a nullptr in this situation, which causes crashes when performing the uv_read_stop call with a nullptr parameter.

Therefore, skip calling uv_read_stop if handle() returns nullptr. This should be fine as there should be no need to call uv_read_stop on a handle that is being cleaned up.