Prevent memory corruption when multiple tasklets perform blocking reads or writes
Prevent the setting of SSL_MODE_RELEASE_BUFFERS,
which is supposed to save about 34k per idle SSL by clearing the buffers used for reading / writing after each operation.

Usually, this is totally fine, as the buffer gets initialized at the start of each operation, and in a standard Python environment these calls are not re-entrant.

However, since there is nothing preventing another tasklet from starting another blocking read / write operation while the first tasklet is blocked, the first tasklet is not able to assume that clearing the buffer is safe.

Therefore, disable this optimization as it could lead to accessing freed memory.