Provide flavored builds of carbon-io
So that logs and telemetry information can be received from non-release builds. Before this change, any flavored run of the monolith environment would pull in the release version of carbon-io, and by extent, the release version of carbon-core. This is, of course, problematic when trying to do logging, or gather telemetry.